Fix _NegateMatcher bug in grepper
Unlike positive matchers that is valid on first match, the exit criteria
is met when all matchers are matched.

Negate matchers must be tested with the whole file and none of negate
matchers can be matched. So it's different logic than testing with
finder. Because finder only test with one line of file path, grepper
needs the whole file.